server/requirements.txt
langchain
langgraph
langchain-core
fastapi[all]
uvicorn[standard]
langchain-community
python-dotenv 
sentence-transformers
pypdf
langchain-google-genai
pymupdf
langchain-ollama
aiosqlite
tavily-python
langchain-tavily
modelcontextprotocol
google-genai
ddgs
httpx
beautifulsoup4
lxml
python-readability
supabase
PyJWT==2.8.0
numpy
pytest
# File processing dependencies
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-multipart>=0.0.6
# HTML sanitization
bleach>=6.0.0
# Math computation
sympy>=1.12
# Additional utilities
aiofiles>=23.2.0
